Add files via upload
I'm thrilled to share that I've successfully completed my first Python project, where I visualized the Diwali sales dataset! 🛍️📊

While this project might be common in the data science community, as a fresher, completing it entirely on my own holds a lot of value for me. It was an incredible learning experience, and I'm proud of the insights I've gained in data visualization, Python, and analysing consumer behaviour.

🔍 **Project Highlights:**
- Analysed Diwali sales data to uncover purchasing patterns and customer behaviour trends.
- Created visualizations using Python (Pandas, Matplotlib, Seaborn) to deliver actionable insights.

This is just the beginning of my journey in data science, and I'm eager to take on more challenging projects. 🚀
I am open to receiving any suggestions that you might find are necessary for enhancing the project and my skills. Hope to receive insights.

Thankful to my teachers for helping me throughout.

#Python #DataScience #DataVisualization #LearningJourney #DiwaliSales #Fresher #Milestone